fix: move type aliases to their respective packages
NavigationStrategy → dimos/mapping/occupancy/types.py
VlModelName → dimos/models/vl/types.py
Remove dimos/core/types.py
