+++ dimos/mapping/occupancy/types.py
@@ -1,0 +1,3 @@
+from typing import Literal, TypeAlias
+
+NavigationStrategy: TypeAlias = Literal["simple", "mixed"]
