+++ dimos/models/vl/types.py
@@ -1,0 +1,3 @@
+from typing import Literal
+
+VlModelName = Literal["qwen", "moondream"]
